Updates   * Added more jsDocs to mutagen lib   * Updated sync task to be mutagen tasks   * Started a create task for mutagen
