+++ src/libs/mutagen/config.js
@@ -1,28 +1,33 @@
 const path = require('path')
+const { deepMerge } = require('jsutils')
 const { loadYml, writeYml, stat } = require('KegFileSys')
 const { generalError } = require('KegUtils/error')
 
-  /**
-  * Helper to build the mutagen config path
-  * <br/> Ensures the config file is saved as a yml file
-  * @function
-  * @param {Object} options - Mutagen Instance options
-  * @param {string} name - Name of mutagen config file
-  *
-  * @returns {Object} - Mutagen config file path
-  */
+/**
+* Helper to build the mutagen config path
+* <br/> Ensures the config file is saved as a yml file
+* @function
+* @param {Object} options - Mutagen Instance options
+* @param {string} name - Name of mutagen config file
+*
+* @returns {Object} - Mutagen config file path
+*/
 const buildPath = ({ configFolder }, name) => {
   name = name.split('.').pop() === '.yml' ? name : `${name}.yml`
   return path.join(configFolder, name)
 }
 
+/**
+ * Default config argument options
+ * @object
+ */
 const configDefs = {}
 
 class Config {
 
   constructor(mutagen){
     this.mutagen = mutagen
-    this.options.sync = deepMerge(configDefs, this.mutagen.options)
+    this.options = deepMerge(configDefs, this.mutagen.options)
   }
 
   /**
@@ -58,7 +63,7 @@ class Config {
   *
   * @returns {boolean} - True if the file exists
   */
-  exists = name => {
+  exists = async name => {
     const [ err, doesExist ] = await stat(buildPath(options, name))
     return err ? generalError(err) : doesExist
   }
