+++ src/tasks/mutagen/stop.js
